+++ src/backend/ci/core/process/api-process/src/main/kotlin/com/tencent/devops/process/api/service/ServicePipelineResource.kt
@@ -446,4 +446,12 @@ interface ServicePipelineResource {
         @PathParam("pipelineId")
         pipelineId: String
     ): Result<PipelineIdInfo?>
+
+    @ApiOperation("根据流水线id获取流水线信息")
+    @GET
+    @Path("/pipelines/{pipelineId}")
+    fun getPipelineNameByIds(
+        @ApiParam("流水线id列表", required = true)
+        pipelineId: String
+    ): Result<SimplePipeline?>?
 }
