+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/api/ServicePipelineResourceImpl.kt
@@ -368,6 +368,14 @@ class ServicePipelineResourceImpl @Autowired constructor(
         return Result(pipelineListFacadeService.getPipelineId(projectCode, pipelineId))
     }
 
+    override fun getPipelineNameByIds(pipelineId: String): Result<SimplePipeline?>? {
+        val pipelineInfos = pipelineListFacadeService.getByPipelineIds(setOf(pipelineId))
+        if (pipelineInfos.isNotEmpty()) {
+            return Result(pipelineInfos[0])
+        }
+        return null
+    }
+
     private fun checkParams(userId: String, projectId: String) {
         checkUserId(userId)
         checkProjectId(projectId)
